Rework flasher as a generic agent to handle all bare metal actions
This is done to have a single service that interfaces with the hardware
to carry out condition related actions. Instead of having one service
for each kind of hardware interaction, now this service does it all
and can be scaled up/down as a NATS worker.

Going ahead we converge the functionalities of Alloy, Flipflop into
this agent. The ctrl library is also included.