internal/ctrl/nats: fix up multiple subject subscription and event processing
This switches to the events.PullOneMsg() which ensures a single message
is fetched and the caller is in better control of which subject
it is pulling from and of the pull failures.

The previous method events.PullMsg() was implicit and would fail on the
first subject pull if there was a timeout.